index.html
Hello, and thanks for grading my project! I don't have a replit subscription but I really need to get this certificate to get my degree at a university course. So I uploaded it to GitHub. It's a bit more complicated than the codes you've been learned at the course but meets all of the requirements. If you open the index.html file, you'll see:

that it's a fully formed page with doctype, head and body elements
that it has way more than three semantic tags.
that it has an h1 tag in the "banner" section
that it has a list and it functions as a navbar
that it has three images in the "services" section
that it has an external link (my existing, real email adrees) in the footer
I hope that'll help you to grade my assesment. Thanks again! Best wishes, Zsolti